Moved code around [skip hazardlib]
